Attempt to fix circular reference issue in test
