+++ src/sas/sascalc/corfunc/corfunc_calculator.py
@@ -13,7 +13,12 @@ from scipy.interpolate import interp1d
 from scipy.signal import argrelextrema
 from numpy.linalg import lstsq
 
-from sas.qtgui.Perspectives.Corfunc.util import TransformedData
+from sas.sascalc.corfunc.calculation_data import (TransformedData,
+                                                  ExtractedParameters,
+                                                  SupplementaryParameters,
+                                                  TangentMethod,
+                                                  LongPeriodMethod)
+
 from sas.sascalc.corfunc.extrapolation_data import ExtrapolationParameters
 
 from sasdata.dataloader.data_info import Data1D
@@ -22,48 +27,6 @@ from sas.sascalc.corfunc.transform_thread import HilbertThread
 from sas.sascalc.corfunc.smoothing import SmoothJoin
 
 
-
-@dataclass
-class SupplementaryParameters:
-    tangent_point_z: float
-    tangent_point_gamma: float
-    tangent_gradient: float
-    first_minimum_z: float
-    first_minimum_gamma: float
-    first_maximum_z: float
-    first_maximum_gamma: float
-    hard_block_z: float
-    hard_block_gamma: float
-    interface_z: float
-    core_z: float
-    z_range: Tuple[float, float]
-    gamma_range: Tuple[float, float]
-
-@dataclass
-class ExtractedParameters:
-    long_period: float
-    interface_thickness: float
-    hard_block_thickness: float
-    soft_block_thickness: float
-    core_thickness: float
-    polydispersity_ryan: float
-    polydispersity_stribeck: float
-    local_crystallinity: float
-
-
-class EntryListEnum(Enum):
-
-    @classmethod
-    def options(cls) -> str:
-        return ", ".join([c.value for c in cls])
-
-class TangentMethod(EntryListEnum):
-    INFLECTION = 'inflection'
-    HALF_MIN = 'half-min'
-
-class LongPeriodMethod(EntryListEnum):
-    MAX = 'max'
-    DOUBLE_MIN = 'double-min'
 
 class CorfuncCalculator:
